backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2

# CV
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4

# ML
torch==2.4.1
torchvision==0.19.1

# LLM
ollama==0.3.3

# Voice
openai-whisper==20231117
pyttsx3==2.90

# Utils
python-dotenv==1.0.1
httpx==0.27.2
